Automated serial port selection
Figured as target demo young, getting them to figure out which serial port to use would be a PITA. Serial initialization now scans through all available ports and assumes whichever one it is receiving data from is the one we want to be using. I'm sure this introduces errors and wouldn't always be a correct assumption but it works for now.